backend/requirements.txt
uvicorn
python-dotenv
fastapi
qdrant-client
google-generativeai